Fixed some formatting and removed instability from graph
