+++ fdm-devito-notebooks/02_wave/src-wave/wave1D/clean.sh
