+++ fdm-devito-notebooks/02_wave/src-wave/wave1D/wave1D_n0_ghost.py
