+++ fdm-devito-notebooks/02_wave/src-wave/wave1D/wave1D_u0.py
@@ -295,3 +295,67 @@ if __name__ == '__main__':
     print('Courant number: %.2f' % C)
     #guitar(C)
     test_convrate_sincos()
+
+# Original Python solver
+def python_solver(I, V, f, c, L, dt, C, T, user_action=None):
+    """Solve u_tt=c^2*u_xx + f on (0,L)x(0,T]."""
+    Nt = int(round(T/dt))
+    t = np.linspace(0, Nt*dt, Nt+1)   # Mesh points in time
+    dx = dt*c/float(C)
+    Nx = int(round(L/dx))
+    x = np.linspace(0, L, Nx+1)       # Mesh points in space
+    C2 = C**2                         # Help variable in the scheme
+    # Make sure dx and dt are compatible with x and t
+    dx = x[1] - x[0]
+    dt = t[1] - t[0]
+
+    if f is None or f == 0 :
+        f = lambda x, t: 0
+    if V is None or V == 0:
+        V = lambda x: 0
+
+    u     = np.zeros(Nx+1)   # Solution array at new time level
+    u_n   = np.zeros(Nx+1)   # Solution at 1 time level back
+    u_nm1 = np.zeros(Nx+1)   # Solution at 2 time levels back
+
+    import time;  t0 = time.perf_counter()  # Measure CPU time
+
+    # Load initial condition into u_n
+    for i in range(0,Nx+1):
+        u_n[i] = I(x[i])
+
+    if user_action is not None:
+        user_action(u_n, x, t, 0)
+
+    # Special formula for first time step
+    n = 0
+    for i in range(1, Nx):
+        u[i] = u_n[i] + dt*V(x[i]) + \
+               0.5*C2*(u_n[i-1] - 2*u_n[i] + u_n[i+1]) + \
+               0.5*dt**2*f(x[i], t[n])
+    u[0] = 0;  u[Nx] = 0
+
+    if user_action is not None:
+        user_action(u, x, t, 1)
+
+    # Switch variables before next step
+    u_nm1[:] = u_n;  u_n[:] = u
+
+    for n in range(1, Nt):
+        # Update all inner points at time t[n+1]
+        for i in range(1, Nx):
+            u[i] = - u_nm1[i] + 2*u_n[i] + \
+                     C2*(u_n[i-1] - 2*u_n[i] + u_n[i+1]) + \
+                     dt**2*f(x[i], t[n])
+
+        # Insert boundary conditions
+        u[0] = 0;  u[Nx] = 0
+        if user_action is not None:
+            if user_action(u, x, t, n+1):
+                break
+
+        # Switch variables before next step
+        u_nm1[:] = u_n;  u_n[:] = u
+
+    cpu_time = time.perf_counter() - t0
+    return u, x, t, cpu_time
