+++ fdm-devito-notebooks/02_wave/src-wave/wave1D/wave1D_u0_s.sh
